feat: [1단계] 인증 로직 리팩터링
