+++ room-escape-web/src/main/java/com/nextstep/web/auth/LoginMemberPrincipal.java
@@ -1,4 +1,4 @@
-package com.nextstep.web.common;
+package com.nextstep.web.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
