+++ room-escape-web/src/main/java/com/nextstep/web/auth/UserDetail.java
@@ -1,0 +1,17 @@
+package com.nextstep.web.auth;
+
+public class UserDetail {
+    private Long id;
+
+    private UserDetail(Long id) {
+        this.id = id;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public static UserDetail from(String id) {
+        return new UserDetail(Long.parseLong(id));
+    }
+}
