+++ room-escape-web/src/main/java/com/nextstep/web/auth/dto/TokenRequest.java
@@ -1,4 +1,4 @@
-package com.nextstep.web.auth;
+package com.nextstep.web.auth.dto;
 
 public class TokenRequest {
     private String username;
