+++ room-escape-web/src/main/java/com/nextstep/web/common/WebConfig.java
@@ -1,6 +1,6 @@
 package com.nextstep.web.common;
 
-import com.nextstep.web.auth.JwtTokenProvider;
+import com.nextstep.web.auth.AuthenticationPrincipalArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
